requirements.txt
zappa
boto3
flask
pillow
slackclient